requirements.txt
tensorflow==1.15.0
opencv-python==4.5.5.62
numpy==1.19.0
Pillow==9.0.1
easydict==1.9
setuptools==60.9.3
PyYAML==6.0
cffi==1.15.0
PyQt5==5.15.6
pip==22.0.3
attrs==21.4.0
wheel==0.37.1
contrib==0.3.0
cloudbiolinux==0.1
cryptography==36.0.1
zipp==3.7.0
matplotlib==3.5.1
scipy==1.4.1
pandas==1.1.5